Clean up the implementation of a LighthouseUser cache; this is mainly used for identifying commentors and filers on tickets.
Since the data doesn't change often (and LH doesn't provide an
updated_at field) it self-updates if it hasn't been updated in 7 days.

LH Users aren't tied to a specific account or project, but rather are
global to the triage app.

With this set of changes, the only code referencing the Lighthouse API
is in models, where it should be.